feat(core): add NoopContext shared test implementation
Centralized no-op Context implementation in basalt-core/testing.rs.
All sub-context methods are stubs returning sensible defaults.
Replaces the duplicated TestContext in basalt-command and prevents
breakage when new trait methods are added.
