+++ crates/basalt-core/src/testing.rs
@@ -1,0 +1,106 @@
+//! Test utilities for crates that need a [`Context`] implementation.
+//!
+//! [`NoopContext`] implements all context traits with no-op methods.
+//! Use it in unit tests that need a `&dyn Context` but don't care
+//! about the actual behavior (e.g., command dispatch tests).
+
+use crate::broadcast::BroadcastMessage;
+use crate::context::{
+    ChatContext, ContainerContext, Context, EntityContext, PlayerContext, WorldContext,
+};
+use crate::gamemode::Gamemode;
+use crate::logger::PluginLogger;
+use basalt_types::{TextComponent, Uuid};
+
+/// A no-op [`Context`] implementation for unit tests.
+///
+/// All methods are stubs that do nothing and return sensible defaults.
+/// The world is a shared memory-only instance (seed 42).
+pub struct NoopContext;
+
+impl PlayerContext for NoopContext {
+    fn uuid(&self) -> Uuid {
+        Uuid::default()
+    }
+    fn entity_id(&self) -> i32 {
+        1
+    }
+    fn username(&self) -> &str {
+        "Steve"
+    }
+    fn yaw(&self) -> f32 {
+        0.0
+    }
+    fn pitch(&self) -> f32 {
+        0.0
+    }
+    fn teleport(&self, _x: f64, _y: f64, _z: f64, _yaw: f32, _pitch: f32) {}
+    fn set_gamemode(&self, _mode: Gamemode) {}
+    fn registered_commands(&self) -> Vec<(String, String)> {
+        Vec::new()
+    }
+}
+
+impl ChatContext for NoopContext {
+    fn send(&self, _text: &str) {}
+    fn send_component(&self, _component: &TextComponent) {}
+    fn action_bar(&self, _text: &str) {}
+    fn broadcast(&self, _text: &str) {}
+    fn broadcast_component(&self, _component: &TextComponent) {}
+}
+
+impl WorldContext for NoopContext {
+    fn world(&self) -> &basalt_world::World {
+        use std::sync::OnceLock;
+        static WORLD: OnceLock<basalt_world::World> = OnceLock::new();
+        WORLD.get_or_init(|| basalt_world::World::new_memory(42))
+    }
+    fn send_block_ack(&self, _sequence: i32) {}
+    fn stream_chunks(&self, _cx: i32, _cz: i32) {}
+    fn persist_chunk(&self, _cx: i32, _cz: i32) {}
+}
+
+impl EntityContext for NoopContext {
+    fn spawn_dropped_item(&self, _x: i32, _y: i32, _z: i32, _item_id: i32, _count: i32) {}
+    fn broadcast_block_change(&self, _x: i32, _y: i32, _z: i32, _block_state: i32) {}
+    #[allow(clippy::too_many_arguments)]
+    fn broadcast_entity_moved(
+        &self,
+        _entity_id: i32,
+        _x: f64,
+        _y: f64,
+        _z: f64,
+        _yaw: f32,
+        _pitch: f32,
+        _on_ground: bool,
+    ) {
+    }
+    fn broadcast_player_joined(&self) {}
+    fn broadcast_player_left(&self) {}
+    fn broadcast_raw(&self, _msg: BroadcastMessage) {}
+}
+
+impl ContainerContext for NoopContext {
+    fn open_chest(&self, _x: i32, _y: i32, _z: i32) {}
+}
+
+impl Context for NoopContext {
+    fn logger(&self) -> PluginLogger {
+        PluginLogger::new("test")
+    }
+    fn player(&self) -> &dyn PlayerContext {
+        self
+    }
+    fn chat(&self) -> &dyn ChatContext {
+        self
+    }
+    fn world_ctx(&self) -> &dyn WorldContext {
+        self
+    }
+    fn entities(&self) -> &dyn EntityContext {
+        self
+    }
+    fn containers(&self) -> &dyn ContainerContext {
+        self
+    }
+}
